python_service/requirements.txt
fastapi==0.115.8
uvicorn==0.24.0
pydantic==2.10.6
python-multipart==0.0.20
websockets==12.0
python-dotenv==1.0.1
browser-use==0.1.36
langchain-openai==0.3.1
